requirements.txt
pandas>=2.2
numpy>=1.26
requests>=2.32
PyYAML>=6.0
pyarrow>=15.0
statsmodels>=0.14
linearmodels>=5.4
matplotlib>=3.9
openpyxl>=3.1
tqdm>=4.66
pytest>=8.0
